fix: Resolve ARM Kobo build errors in reader modules
- reader_annotations_ext.rs: Add PartialEq derive to Annotation struct
- reader_rendering_ext.rs: Fix type mismatches in arithmetic operations
- reader_rendering_ext.rs: Fix draw_pixmap call to use Point instead of Rectangle
- reader_rendering_ext.rs: Change render_page return type to anyhow::Error
- reader.rs: Add CycleDir import for go_to_results_neighbor method
- reader.rs: Remove duplicate event handler method definitions
- document/mod.rs: Add page and level fields to TocEntry struct

Build status: 53 errors remaining (down from 158)
Most remaining errors are in ui_toggles modules

+++ crates/core/src/document/html/engine_text/text_renderer.rs
@@ -237,8 +237,8 @@ pub mod utils {
         let width = text.len() as f32 * font_size * 0.6;
         let height = font_size;
         Rectangle::new(
-            crate::geom::Point::new(0.0, 0.0),
-            crate::geom::Point::new(width, height),
+            crate::geom::Point::new(0, 0),
+            crate::geom::Point::new(width as i32, height as i32),
         )
     }
